♻️move error to domain
it had dependency on domain
