+++ packages/core/src/domain/console/consoleObservable.ts
@@ -1,5 +1,5 @@
 import { computeStackTrace } from '../tracekit'
-import { createHandlingStack, formatErrorMessage, toStackTraceString } from '../../tools/error'
+import { createHandlingStack, formatErrorMessage, toStackTraceString } from '../error/error'
 import { mergeObservables, Observable } from '../../tools/observable'
 import { find, jsonStringify } from '../../tools/utils'
 import { ConsoleApiName } from '../../tools/display'
