+++ packages/core/src/domain/error/error.ts
@@ -1,10 +1,10 @@
-import type { StackTrace } from '../domain/tracekit'
-import { computeStackTrace } from '../domain/tracekit'
-import { ExperimentalFeature, isExperimentalFeatureEnabled } from './experimentalFeatures'
-import { callMonitored } from './monitor'
-import { sanitize } from './sanitize'
-import type { ClocksState } from './timeUtils'
-import { jsonStringify, noop } from './utils'
+import type { StackTrace } from '../tracekit'
+import { computeStackTrace } from '../tracekit'
+import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../../tools/experimentalFeatures'
+import { callMonitored } from '../../tools/monitor'
+import { sanitize } from '../../tools/sanitize'
+import type { ClocksState } from '../../tools/timeUtils'
+import { jsonStringify, noop } from '../../tools/utils'
 
 export const NO_ERROR_STACK_PRESENT_MESSAGE = 'No stack, consider using an instance of Error'
 export const PROVIDED_ERROR_MESSAGE_PREFIX = 'Provided'
