+++ packages/core/src/domain/error/trackRuntimeError.ts
@@ -1,8 +1,8 @@
-import type { RawError } from '../../tools/error'
-import { ErrorSource, computeRawError, ErrorHandling } from '../../tools/error'
 import type { Observable } from '../../tools/observable'
 import { clocksNow } from '../../tools/timeUtils'
 import { startUnhandledErrorCollection } from '../tracekit'
+import { ErrorSource, computeRawError, ErrorHandling } from './error'
+import type { RawError } from './error'
 
 export function trackRuntimeError(errorObservable: Observable<RawError>) {
   return startUnhandledErrorCollection((stackTrace, originalError) => {
