+++ packages/core/src/index.ts
@@ -79,7 +79,7 @@ export {
   getFileFromStackTraceString,
   NO_ERROR_STACK_PRESENT_MESSAGE,
   PROVIDED_ERROR_MESSAGE_PREFIX,
-} from './tools/error'
+} from './domain/error/error'
 export { Context, ContextArray, ContextValue } from './tools/context'
 export { areCookiesAuthorized, getCookie, setCookie, deleteCookie, COOKIE_ACCESS_DELAY } from './browser/cookie'
 export { initXhrObservable, XhrCompleteContext, XhrStartContext } from './browser/xhrObservable'
