+++ packages/core/src/transport/httpRequest.ts
@@ -2,7 +2,7 @@ import type { EndpointBuilder } from '../domain/configuration'
 import { addTelemetryError } from '../domain/telemetry'
 import type { Context } from '../tools/context'
 import { monitor } from '../tools/monitor'
-import type { RawError } from '../tools/error'
+import type { RawError } from '../domain/error/error'
 import { newRetryState, sendWithRetryStrategy } from './sendWithRetryStrategy'
 import type { FlushReason } from './batch'
