+++ packages/core/src/transport/startBatchWithReplica.ts
@@ -1,5 +1,5 @@
 import type { Configuration, EndpointBuilder } from '../domain/configuration'
-import type { RawError } from '../tools/error'
+import type { RawError } from '../domain/error/error'
 import type { Context } from '../tools/context'
 import type { Observable } from '../tools/observable'
 import type { PageExitEvent } from '../browser/pageExitObservable'
